pyproject.toml
[tool.poetry]
name = "semester_1"
version = "0.1.0"
description = ""
authors = ["Giknyx <bexultanovsultan@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Flask = "^2.0.2"
Flask-Migrate = "^3.1.0"
Flask-SQLAlchemy = "^2.5.1"
psycopg2-binary = "^2.9.1"
Flask-Login = "^0.5.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
